+++ example/package.json
@@ -15,7 +15,6 @@
     "color": "^0.11.4",
     "escape-string-regexp": "^1.0.5",
     "expo": "~21.0.2",
-    "lodash.merge": "^4.6.0",    
     "prop-types": "^15.5.10",
     "react": "16.0.0-alpha.12",
     "react-native": "~0.48.4",
