+++ example/src/ExampleList.js
@@ -2,8 +2,9 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { ListView, Text, StyleSheet, View } from 'react-native';
-import { TouchableRipple, Divider, withTheme } from 'react-native-paper';
+import { ListView } from 'react-native';
+import { Divider, withTheme } from 'react-native-paper';
+import ExampleListRow from './ExampleListRow';
 import ButtonExample from './ButtonExample';
 import FABExample from './FABExample';
 import CardExample from './CardExample';
@@ -49,54 +50,26 @@ class ExampleList extends Component {
     navigation: PropTypes.object,
   };
 
-  state = {
-    value: 0,
-  };
-
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.theme !== this.props.theme) {
-      /// Super dirty way to update the listview when now rows change
-      this.setState({ value: this.state.value + 1 });
-    }
-  }
-
-  _renderRow = id => {
-    const { theme: { colors: { paper, text } } } = this.props;
-    return (
-      <TouchableRipple
-        style={[styles.item, { backgroundColor: paper }]}
-        onPress={() => this.props.navigation.navigate(id)}
-      >
-        <Text style={[styles.text, { color: text }]}>{examples[id].title}</Text>
-      </TouchableRipple>
-    );
-  };
+  _renderRow = id => (
+    <ExampleListRow
+      title={examples[id].title}
+      onPress={() => this.props.navigation.navigate(id)}
+    />
+  );
 
   _renderSeparator = (sectionId, rowId) => <Divider key={rowId} />;
 
   render() {
     const { theme: { colors: { background } } } = this.props;
     return (
-      <View style={{ flex: 1, backgroundColor: background }}>
-        <ListView
-          key={this.state.value}
-          dataSource={dataSource}
-          renderRow={this._renderRow}
-          renderSeparator={this._renderSeparator}
-        />
-      </View>
+      <ListView
+        contentContainerStyle={{ flex: 1, backgroundColor: background }}
+        dataSource={dataSource}
+        renderRow={this._renderRow}
+        renderSeparator={this._renderSeparator}
+      />
     );
   }
 }
-
-const styles = StyleSheet.create({
-  item: {
-    padding: 16,
-  },
-  text: {
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-});
 
 export default withTheme(ExampleList);
